Liip imagine installation

+++ src/Controller/Admin/ArticleController.php
@@ -3,10 +3,12 @@
 namespace App\Controller\Admin;
 
 use App\Entity\Article;
+use App\Entity\Image;
 use App\Form\ArticleType;
 use App\Repository\ArticleRepository;
 use Symfony\Bundle\FrameworkBundle\Controller\AbstractController;
 use Symfony\Component\Form\Extension\Core\Type\SubmitType;
+use Symfony\Component\HttpFoundation\JsonResponse;
 use Symfony\Component\HttpFoundation\Request;
 use Symfony\Component\HttpFoundation\Response;
 use Symfony\Component\Routing\Annotation\Route;
@@ -39,7 +41,25 @@ class ArticleController extends AbstractController
         $articleForm->handleRequest($request);
 
         if ($articleForm->isSubmitted() && $articleForm->isValid()) {
-            
+
+            // On récupère les images transmisent
+            $images = $articleForm->get('images')->getData();
+            foreach ($images as $image) {
+                //On génére un nouveau nom pour chaque image
+                $fichier = md5(uniqid()) . '.' . $image->guessExtension();
+
+                // On copie le fichier dans le dossier Uploads/images
+                $image->move(
+                    $this->getParameter('images_directory'),
+                    $fichier
+                );
+
+                // On stock l'image dans la base de donnée
+                $img = new Image;
+                $img->setName($fichier);
+                $article->addImage($img);
+            }
+
             $em = $this->getDoctrine()->getManager();
             $em->persist($article);
             $em->flush();
@@ -47,18 +67,17 @@ class ArticleController extends AbstractController
             $this->addFlash('success', 'Un article a été ajouté !');
 
             return $this->redirect($this->generateUrl('app_admin_article_home'));
-            
         }
 
 
-        
+
         return $this->render('admin/admin_article/insert.html.twig', [
             'current_admin_menu' => 'app_admin_article',
             'articleForm' => $articleForm->createView(),
         ]);
     }
 
-    #[Route('/edit/{id}', name: 'edit')]    
+    #[Route('/edit/{id}', name: 'edit')]
     public function edit(Request $request, Article $article): Response
     {
 
@@ -70,7 +89,25 @@ class ArticleController extends AbstractController
         $articleForm->handleRequest($request);
 
         if ($articleForm->isSubmitted() && $articleForm->isValid()) {
-            
+
+            // On récupère les images transmisent
+            $images = $articleForm->get('images')->getData();
+            foreach ($images as $image) {
+                //On génére un nouveau nom pour chaque image
+                $fichier = md5(uniqid()) . '.' . $image->guessExtension();
+
+                // On copie le fichier dans le dossier Uploads/images
+                $image->move(
+                    $this->getParameter('images_directory'),
+                    $fichier
+                );
+
+                // On stock l'image dans la base de donnée
+                $img = new Image;
+                $img->setName($fichier);
+                $article->addImage($img);
+            }
+
             $em = $this->getDoctrine()->getManager();
             $em->persist($article);
             $em->flush();
@@ -78,39 +115,87 @@ class ArticleController extends AbstractController
             $this->addFlash('success', 'L\'article a été modifié !');
 
             return $this->redirect($this->generateUrl('app_admin_article_home'));
-            
         }
 
 
-        
+
         return $this->render('admin/admin_article/edit.html.twig', [
             'current_admin_menu' => 'app_admin_article',
             'articleForm' => $articleForm->createView(),
+            'article' => $article,
         ]);
     }
 
-    #[Route('/remove/{id}', name: 'remove')]    
+    #[Route('/remove/{id}', name: 'remove')]
     public function remove(Request $request, Article $article): Response
     {
- 
-        $em = $this->getDoctrine()->getManager();
-        $em->remove($article);
-        $em->flush();
 
-        $this->addFlash('success', 'L\'article a été supprimé avec succès !');
+        if ($this->isCsrfTokenValid('delete' . $article->getId(), $request->get('_token'))){
 
-        return $this->redirect($this->generateUrl('app_admin_article_home'));
+            $images = $article->getImages();
+            foreach($images as $image) {
+
+                $nom = $image->getName();
+                unlink($this->getParameter('images_directory') . '/' . $nom);
+    
+                $em = $this->getDoctrine()->getManager();
+                $em->remove($image);
+
+                $this->addFlash('success', 'Les images de l\'article ont été supprimées avec succès !');
+            }
+
+            $em = $this->getDoctrine()->getManager();
+            $em->remove($article);
+            $em->flush();
+    
+            $this->addFlash('success', 'L\'article a été supprimé avec succès !');
+    
+            return $this->redirect($this->generateUrl('app_admin_article_home'));
+
+        } else {
+
+            $this->addFlash('error', 'Le token est invalide');
+
+            return $this->redirect($this->generateUrl('app_admin_article_home'), 302);
+
+        }
+
 
     }
 
     #[Route('/setActive/{id}', name: 'SetActive')]
-    public function setActive(Article $article): Response
+    public function setActive(Article $article): void
     {
-        $article->setIsActive(($article->getIsActive())?false:true);
+        $article->setIsActive(($article->getIsActive()) ? false : true);
         $em = $this->getDoctrine()->getManager();
         $em->persist($article);
         $em->flush();
+    }
 
-        return new Response("true");
+    #[Route("/remove/image/{id}", name: "remove_image", methods: ['DELETE'])]
+    public function deleteImage(Image $image, Request $request)
+    {
+
+        //On vérifie si le token est valide
+        if ($this->isCsrfTokenValid('delete' . $image->getId(), $request->get('_token'))) {
+
+            $nom = $image->getName();
+            unlink($this->getParameter('images_directory') . '/' . $nom);
+
+            $em = $this->getDoctrine()->getManager();
+            $em->remove($image);
+            $em->flush();
+
+            return $this->redirect($this->generateUrl('app_admin_article_edit', [
+                'id' => $image->getArticle()->getId()
+            ]));
+
+        } else {
+
+            $this->addFlash('error', 'Le token est invalide');
+
+            return $this->redirect($this->generateUrl('app_admin_article_home'), 302);
+
+        }
     }
 }
